+++ test/Integration/Dialect/TOSA/i8xi8_mul_elem_16x1024_broadcast_1024/i8xi8_mul_elem_2d_broadcast_2d.mlir
@@ -12,7 +12,7 @@
 
 module {
   func.func @dut(%arg0: tensor<16x1024xi8>, %arg1: tensor<1x1024xi8>) -> (tensor<16x1024xi8>) {
-    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i32} : (tensor<16x1024xi8>, tensor<1x1024xi8>)  -> (tensor<16x1024xi8>)
+    %1 = "tosa.mul"(%arg0,%arg1) {shift = 0 : i8} : (tensor<16x1024xi8>, tensor<1x1024xi8>)  -> (tensor<16x1024xi8>)
     return %1 : tensor<16x1024xi8>
   }
 }
